Il a cassé game j'y crois pas

+++ code/game.py
@@ -2,17 +2,13 @@ import socket, pickle, sys, os
 from deckmanager import DeckManager
 from player import Player
 
-import time
-
 DEBUG = False
 
 SOCKET = 0
 PLAYER = 1
 LIFE = 10
 
-DECK1 = "White"
-DECK2 = "Black"
-DECKTEST = "Test_Tristan"
+DEFAULT_DECK = "Test_Tristan"
 
 class Game:
 
@@ -22,6 +18,7 @@ class Game:
 		self.__deckmanager = DeckManager()
 		self.__slots = slots
 		self.__players = []
+		self.__dead_players = []
 
 	def get_socket(self):
 		return self.__socket
@@ -160,7 +157,7 @@ class Game:
 		for player in self.__players:
 	
 			# Création automatique du deck sans demander au client
-			self.__deckmanager.add(DECKTEST)
+			self.__deckmanager.add(DEFAULT_DECK)
 			deck = self.__deckmanager.copy_deck(0)
 
 			# Création de l'objet Player en lui passant le deck
@@ -186,6 +183,21 @@ class Game:
 			# Exécution de la phase
 			print("[PLAYER " + str(player[PLAYER].get_id()+1) + "] Phase Mulligan")
 			self.mulligan(player[PLAYER].get_id())
+
+	def concede(self, index):
+		# Envoi vers le client : Acceptation (10.2.1)
+		self.send_signal(index,"ACCEPT")
+		
+		# On met la vie du joueur a 0
+		self.__players[index][PLAYER].set_life(0)
+
+		# Envoi vers le client : Etat de la partie (10.2.2)
+		self.send_gamestate(index)
+		
+		# Envoi vers le client : Signal de mort (23)
+		self.send_signal(index,"DEATH")
+		
+		self.__dead_players.append(index)
 
 	def mulligan(self, index):
 
@@ -238,6 +250,12 @@ class Game:
 				self.send_gamestate(index)
 
 				break
+			
+			elif(data.get("type") == "CONCEDE"):
+
+				self.concede(index)
+
+				break
 
 			else:
 
@@ -251,7 +269,7 @@ class Game:
 		data = None
 
 		# Rafraichissement de l'écran
-		self.clear_terminal()	
+		self.clear_terminal()
 
 		while self.players_alive() > 1:
 
@@ -320,6 +338,7 @@ class Game:
 
 						# Envoi vers le client : Signal de mort (23)
 						self.send_signal(player[PLAYER].get_id(),"DEATH")
+						self.__dead_players.append(player[PLAYER].get_id())
 
 		# Envoi des résultats
 		for player in self.__players:
@@ -329,10 +348,11 @@ class Game:
 				# Envoi vers le client : Signal de mort (24.1)
 				self.send_signal(player[PLAYER].get_id(),"VICTORY")
 
-			else:
+			elif player[PLAYER].get_id() not in self.__dead_players:
 				
 				# Envoi vers le client : Signal de mort (24.2)
 				self.send_signal(player[PLAYER].get_id(),"DEATH")
+				self.__dead_players.append(player[PLAYER].get_id())
 
 	def effect_phase(self, index):
 
@@ -363,6 +383,12 @@ class Game:
 
 				# Envoi vers le client : Etat de la partie (7.2.2)
 				self.send_gamestate(index)
+
+				break
+			
+			elif(data.get("type") == "CONCEDE"):
+
+				self.concede(index)
 
 				break
 
@@ -402,6 +428,12 @@ class Game:
 				self.send_gamestate(index)
 
 				break
+			
+			elif(data.get("type") == "CONCEDE"):
+
+				self.concede(index)
+
+				break
 
 			else:
 
@@ -428,6 +460,12 @@ class Game:
 
 				# Envoi vers le client : Etat de la partie (13.1.2)
 				self.send_gamestate(index)
+
+				break
+			
+			elif(data.get("type") == "CONCEDE"):
+
+				self.concede(index)
 
 				break
 
@@ -467,8 +505,20 @@ class Game:
 				self.send_gamestate(index)
 
 				break
+				
+			elif(data.get("type") == "CONCEDE"):
+
+				self.concede(index)
+
+				break
 
 			else:
 
 				# Envoi vers le client : Refus (19.3)
 				self.send_signal(index,"DECLINE")
+				
+				
+				
+				
+				
+
